WM: don't skip AutoSave if modal operator is 'fileselect'
AutoSave is skipped if we have a file select handle open.

But a user could inadvertently minimize the window from that handle,
causing AutoSave to no longer work.

So skip this case of skipping AutoSave.

---
That's a shot in the dark.

Commit cb8f7fd385 implemented the following behavior for AutoSave:
- Autosave is postponed by 10 seconds when a modal operator is running, e.g. transform or file browsing.

But it's not clear why or which state is not safe for AutoSave.

Those conditions that skip AutoSave could be causing inconveniences that are hard to track down. It is possible for example that it is the cause commented on https://projects.blender.org/blender/blender/issues/107421#issuecomment-931721

That's why this PR proposes to be a little less strict for cases of skipping  AutoSave.