Vulkan: Make ShaderInput optional
ShaderInput lookup used to be explicit as in we know for sure that the
shader input exists. This was guarded by a assert. During development
of the graphics pipeline we came to the conclusion that this isn't
always the case due to its late bindings of resources.

This PR Makes the shader input lookup return an optional result.
This result can then be checked depending on the area where it is used
to validate the existence.

NOTE: Unclear if this is still needed and has to be re-evaluated.
So question is if we are still able to reproduce the behavior that
requires this. It could be solved by the late binding of resources and
this was just a temp work-around to come to the late binding solution.