Vulkan: Swap chain
This PR replaces the previous implementation of the swap chain.
The previous implementation was based on a gaming loop where
inside a single function an image is requested, the drawing occurs
and the requested image is presented on screen.

In blender the drawing isn't controlled in a single function and
this approach lead to freezes, missing frames and other
artifacts.

This approach might not be the final approach but a step into a
direction where the acquiring of the next image in the swap chain
is separated from the swapping.