Cleanup: Avoid state creation for first simulation input node evaluation
On an evaluation of the simulation input node when there is no previous
state, the node's inputs are evaluated and passed directly to the
outputs. Currently this a bit confusing, since they are first converted
into a simulation state and _then_ passed to the outputs. This commit
simplifies the node to avoid creating a simulation state, just moving
the values directly instead. The inputs are also set to unused when
there is a previous state too.