Geometry Nodes: Add 'Fill Interior' option to Mesh to SDF Volume
Adds an input socket 'Fill Interior to the Mesh to SDF Volume node.
This will make the interior band fill the entire interior of the mesh.
This is useful for example to be able to use the Threshold of
'Volume to Mesh' smoothly all the way down until the mesh disappears
without needing to also increase the exterior band width.