macOS Metal viewport: Blocking artefacts on selected faces
Previous fix 22e4e973104 worked on low density screens, but failed
when using hi-density or hi resolution. This PR increases the
offset to check for better results.

NOTE: when the numbers have been found we should check if we
need to make the offset scale based on the viewport resolution.

Ref: #105830