USD export: option to add a root prim.
Added a root_prim_path USD string option.  If this option is set to
a prim path (e.g., '/World'), a transform primitive with the given
path will be added to the stage as the parent of all exported
primitives. This will allow transforming the scene by this root
node in other applications and is also required for loading the
exported scene as a reference in other USDs.

Also added logic to include the material library node, which contains
all exported materials, under this root prim.

Finally, added an export_params_valid() function to verify that the
root_prim_path is valid before starting the export jog.  This function
can be extended in the future to perform additional validation of
other export parameters.