WIP: CMake: plumbing for modern CMake usage
This is the minimal change required to start using modern CMake in the
blender build system. Previous iterations of this patch did more work
like completely transitioning the guarded alloc and atomic libs to
modern cmake, this however came at a great cost of merge conflicts
and is one of the reasons this has dragged on for literally years now
as the time required for me to keep a patch in sync with main just isn't
there.

This change is designed to allow small incremental changes to the
build system rather than doing it in one big bang which would be
unmaintainable (for me)

The biggest functional change is, previously all libraries in the LIB
section of a `blender_add_test_lib` had the `INTERFACE` scope, which
is rarely, if ever the correct scope. This diff changes this to
`PRIVATE`

Concrete implications of this diff

Right now the only library using any modern cmake is `bf_intern_atomic`
which is an header only interface library that will just advertise its include
directories.

This allows us to clean up any CMakeLists.txt that adds `../../../intern/atomic`
to its `INC` section to remove it in `INC`, add a `PRIVATE bf_intern_atomic` to
`LIB`

Ideally when all targets are modern targets fudging things in `INC` will be a
thing of the past. This also means that all dependencies between targets will
be properly expressed at that time, since right now there are quite a few
occasions of questionable behaviour, ie compare the `INC` section of
`source\blender\editors\space_view3d\CMakeLists.txt` to its `LIB` section.
(not shaming `space_view3d` here, it's the rule rather than the exception)

Where to next

I'd like to land these core changes as soon as the bcon1 window for 4.0 opens
then do further incremental clean-ups with the obvious first choices being
the atomic and guardedalloc libraries.