Light Linking: Support shadow linking to emissive meshes
This includes picking up a mesh emitter with the check for a fully
blocked light by a blocker with default linking configuration.

Implemented using an intersection loop, which might not have the
best performance compared to intersect-all behavior, but allows to
use some specialized for the shadow linking ray manipulations and
object flags check.

Ref #104972