Light Linking: Unify meaning of the light sets
Previously the default (affect all) light set would depend on
whether there is light linking in the scene or not: if there is
light linking the set would be set to 0, but when there is light
linking and light is not linked to anything the set would be all
bits set.

This change makes it so the light set is always all bits sets
regardless of whether there is no light linking in the scene at
all, or the light is not explicitly linked to anything.

This simplifies the mental model while working on the kernel
code and should not cause any functional changes. Additionally,
it seems that it will allow to avoid some of the kernel_features
check in the kernel prior to access the bitmask.