OK Color initial implementation
The foundation is there, with a proper external module and build option.
Björn Ottosson's original code is a CPP header; I ported it to C with a header that only exposes the desired functions. I might add a diff for this later to keep track of changes.
The actual integration of these functions into Blender's color math is hacky and doesn't work properly, which is what I intend to work on next.