Make ok_color functions similar to math_color
You no longer need to use the structs to use these functions.
The clamping and NaN checks are done in the new functions, to keep the rest of the code closer to the original.
The actual calling of these functions is still done in math_color.c, which is definitely the wrong place; one step at a time.