Mesh: Remove face map list, convert to integer attribute
See #105317

Face maps were added as a prototype of a new rigging solution during
2.8 development. Their storage is redundant with the newer generic
attribute system (specifically with integer face attributes), and
they were never used much. This removes the face map list and
converts the storage to an attribute with the name `face_maps`.
There is nowehere to store the face map names anymore, so those
are not converted.

If it's preferred, the two features removed here (gizmo drawing
and edit mode select similar) could be reimplemented for integer
attributes). Personally I think it's better to wait for more
generic attribute editing/gizmo features though.