UI: Edit Mode Options panel cleanup
This is an alternative to the work done in 58a75d01a6

In that commit the AutoMerge settings were separated into a Sub-Panel which is dependent on the `Options` Panel.

However AutoMerge is a property of transform operations, so it seems more convenient to have this option bundled with transform settings.

Another thing is that the `Live Unwrap` option has no real relation to Transform.

Therefore, the proposal is to create 2 Sub-Panels for the Options Panel:
- Transform
- UVs

To avoid creating even more hierarchy, `AutoMerge` settings are no longer separated into an isolated Panel.

And `Live Unwrap` is now part of the UVs Sub-Panel.