UI: move Placement Settings to the Options popover
Since a generic snap cursor was implemented (which can be used by Tools and by DragDrop), the Placement Settings are no longer properties of the "VIEW3D_OT_interactive_add" operator.

They now are part of the `ToolSettings` and can be used for:
- Drag and Drop
- Measure Tool
- Placement Tool

This makes the interface confusing, as the user doesn't know where to edit the Placement Settings for Drag and Drop.

The solution proposed in this PR is to move the Placement Settings to the Options popover (top right hand corner).