Mesh: Move vertex/edge crease to generic attributes
Store subdivion surface creases in two new named float attributes:
- `crease_vert`
- `crease_edge`

The attributes are naming conventions, so their data type and domain
aren't enforced, and may be interpolated when necessary. Editing tools
and the subdivision surface modifier use the hard-coded name. It might
be best if these were edited as generic attributes in the future, but
in the meantime using generic attributes helps.

The attributes are visible in the list, which is how they're now meant
to be removed. They are now interchangeable with any tool that works
with the generic attribute system-- even tools like vertex paint can
affect creases now.

This is a breaking change. Forward compatibility isn't preserved, and
the `.vertex_creases` and `.edge_creases` and `crease` properties in
RNA are removed in favor of making a smaller API surface area with just
the attribute API.

A few extrude node test files have to be updated because of different
attribute interpolation behavior.