Fix #108096: UI: copy-paste array buttons can change source properties
Pasting numerical array buttons happens with `Ctrl + Alt + V`.
Holding `Alt` also triggers uiSelectContext, so having other nodes/
objects etc. selected while doing this would try to copy the pasted
values back to other objects (possibly to the ones you pasted from) and
that happens relative to the original value, so the value actually
changes.

To prevent the "copy-to-selected" behavior, check for the specific button
combination in `ui_apply_but` and skip it in that case.