WIP: BLI: Improve IndexMask::complement() performance
Instead of reusing `from_predicate` and lookup in the source mask,
scan the mask once, inserting segments between the original indices.

Theoretically this improves the performance from O(N*log(N)) to O(N).
But with the small constant offset of the former, the improvement won't
be that nice.

TODO:
- More performance testing. I didn't see much change in the test code runtime.